fix(handler) improve error formatting for invalid templates
Despite the validation on the schema, it is still possible to write
invalid templates which will result on a 500 error in runtime.

The current result will provoke that the `current_template` variable
will be `nil`. So the error resulting from this is a 500 `nil-variable`
error.

This change makes it a more explicit 500 error, printing the error
message returned by trying to compile the template in the logs.

It changes nothing for the end user, but it makes it easier to spot a
possible bug while configuring the plugin.