tests(handler) fix failing test
In EE 2.1.x + Cassandra, the pattern appeared 4 times, in 2 groups of 2,
fairly separated.

I suspect EE + Cassandra is keeping logs from previous specs, so I am
now counting the starting appearances and using that to compare.

I have also replaced the  == 1 or == 2 test with a >=1, because I
believe the test should not fail if we suddenly encounter a 3 (maybe
introduced by a different test that took a while to reach the log)